Add files via upload
Added lockString option to show and hide commands.  This helps with modules that show and hide themselves for various reasons.  Combined with a module that set its own lockstrings when showing and hiding, it makes sure to show only when both MMM-Facial-Recognition wants to show and the native module wants to show.